// packages/twenty-server/src/database/commands/upgrade-version-command/2-31/types/record-page-reown-updates.type.ts
export type RecordPageReownUpdate = {
  id: string;
  update: {
    universalIdentifier?: string;
    isSystemSideEffect?: boolean;
    applicationId?: string;
  };
};

export type RecordPageReownUpdates = {
  pageLayoutUpdates: RecordPageReownUpdate[];
  pageLayoutTabUpdates: RecordPageReownUpdate[];
  pageLayoutWidgetUpdates: RecordPageReownUpdate[];
  viewUpdates: RecordPageReownUpdate[];
  viewFieldUpdates: RecordPageReownUpdate[];
  viewFieldGroupUpdates: RecordPageReownUpdate[];
};